refactor(a380x): move PRIM publishers to shared
