+++ fbw-a380x/src/systems/shared/src/publishers/PrimFctlPublisher.ts
